:truck: Rename ToursGrid to CardGrid
