+++ client/src/components/helpers/CardContainer.js
@@ -3,7 +3,7 @@ import styled from "styled-components/macro";
 const CardContainer = styled.div`
   position: relative;
 
-  width: clamp(300px, 10vw, 350px);
+  width: clamp(320px, 20vw, 400px);
   margin: auto;
   padding: 1rem;
   text-align: center;
